Rename DemoSignMessage to Demo. Extract sign message page logic out into SignMessagePage from SignMessagePanel
